+++ modules/profile/screens/ProfileScreen.tsx
@@ -1,44 +1,47 @@
-import ThemeToggle from '@/components/ThemeToggle';
-import { useAuthStore } from '@/stores/useAuthStore';
-import { getBiometricEnabled } from '@/utils/biometricFlag';
-import { appColors } from '@/utils/colors';
-import { MaterialIcons } from '@expo/vector-icons';
-import { useState } from 'react';
-import { ScrollView, Text, TouchableOpacity, View } from 'react-native';
-import BiometricToggle from '../components/BiometricView';
+import ThemeToggle from "@/components/ThemeToggle";
+import { useAuthStore } from "@/stores/useAuthStore";
+import { getBiometricEnabled } from "@/utils/biometricFlag";
+import { appColors } from "@/utils/colors";
+import { MaterialIcons } from "@expo/vector-icons";
+import { useState } from "react";
+import { ScrollView, Text, TouchableOpacity, View } from "react-native";
+import BiometricToggle from "../components/BiometricView";
 
 export default function ProfileScreen() {
-  const { session, role,token, signOut, signOutSoft } = useAuthStore();
-  const [isloading, setIsLoading] = useState<boolean>(false)
-
+  const { session, role, token, name, signOut, signOutSoft } = useAuthStore();
+  const [isloading, setIsLoading] = useState<boolean>(false);
 
   const handleSignOut = async () => {
-    setIsLoading(true)
+    setIsLoading(true);
     try {
       const isEnabled = await getBiometricEnabled();
 
       if (isEnabled) {
         await signOutSoft();
-      }
-      else {
+      } else {
         await signOut();
       }
-    }
-    catch (error: unknown) {
-      const errorMessage = error instanceof Error ? error.message : String(error);
+    } catch (error: unknown) {
+      const errorMessage =
+        error instanceof Error ? error.message : String(error);
       alert("Error " + errorMessage);
-    }
-    finally {
+    } finally {
       setIsLoading(false);
     }
-
   };
 
   return (
-    <ScrollView className="flex-1 p-6 bg-background dark:bg-dark-background" alwaysBounceVertical={true} indicatorStyle={'black'}>
-      <View className='items-center gap-4 pb-60'>
-
-
+    <ScrollView
+      className="flex-1 p-6 bg-background dark:bg-dark-background"
+      alwaysBounceVertical={true}
+      indicatorStyle={"black"}
+    >
+      <View className="items-center gap-4 pb-60">
+        {name && (
+          <View className="items-center px-4 py-3 rounded-xl w-[80%]   bg-componentbg dark:bg-dark-componentbg">
+            <Text className="text-lg  dark:text-white">{name}</Text>
+          </View>
+        )} 
         {session?.user?.email && (
           <View className="items-center px-4 py-3 rounded-xl w-[80%]   bg-componentbg dark:bg-dark-componentbg">
             <Text className="text-lg  dark:text-white">
@@ -47,14 +50,6 @@ export default function ProfileScreen() {
           </View>
         )}
 
-        {/* {role && (
-          <View className="items-center px-4 py-3 rounded-xl w-[80%]   bg-componentbg dark:bg-dark-componentbg">
-
-            <Text className="text-lg  dark:text-white">
-              Rol: {role}
-            </Text>
-          </View>
-        )} */}
         {/* {token && (
           <View className="items-center px-4 py-3 rounded-xl w-[80%]   bg-componentbg dark:bg-dark-componentbg">
 
@@ -64,10 +59,8 @@ export default function ProfileScreen() {
           </View>
         )} */}
 
-
         <BiometricToggle />
         <ThemeToggle />
-
 
         <TouchableOpacity
           className="flex-row items-center bg-error/10 p-3 rounded-lg border border-error/20 "
@@ -79,7 +72,6 @@ export default function ProfileScreen() {
           <Text className="text-error font-medium ml-2">Salir</Text>
         </TouchableOpacity>
       </View>
-
     </ScrollView>
   );
 }