+++ CHANGELOG.md
@@ -5,25 +5,28 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [Unreleased]
-
 ## [0.0.3]
 
 ### Added
 
 -ui component `CustomImage`
+-Name in `useAuthStore` from supabase
 
 ### Chaged
 
 - UI and styles in TitleText from CustomFlatList
 - Text for list in OrdersModule
 - buttons in HomeScreen
+- function in  `useHomeScreen` for get the current mont and year
+- hide the emojis and icon in home screen and drawer
 
 ### Fixed
 
 - The app crashes when the `pedido.co_art` is null, a validation check has been implemented.
 - Added validation for discount for product list in Orders Module `hasDiscount = discount !== "0" && discount.length > 0`
 - Expanding the second item and image in ProductList causes the first item to render incompletely. Only one item was configured for expansion.
+- Prevented automatic write on first render: Added a didMount flag to avoid triggering `setBiometricEnabled()` when th  e component first loads im profile.
+- The biometric toggle no longer defaults to true on first app launch if the saved value is false.
 
 ## [0.0.2](https://github.com/JavierTaborda/AppFrigilux/releases/tag/v0.0.2) - 2025-10-08